requirements.txt
playwright==1.48.0
gspread==6.1.4
google-auth==2.35.0
python-dotenv==1.0.1
tenacity==9.0.0
